Refactor file upload modules
